requirements.txt
litellm[proxy]
PyYAML>=6,<7
pytest>=8,<9
